refactor: test에 truncate.sql 적용
