+++ src/test/java/wooteco/subway/acceptance/AcceptanceTest.java
@@ -1,4 +1,4 @@
-package wooteco.subway;
+package wooteco.subway.acceptance;
 
 import io.restassured.RestAssured;
 import io.restassured.specification.RequestSpecification;
