+++ src/test/java/wooteco/subway/acceptance/line/LineStationAcceptanceTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Test;
 import wooteco.subway.service.line.dto.LineDetailResponse;
 import wooteco.subway.service.line.dto.LineResponse;
 import wooteco.subway.service.station.dto.StationResponse;
-import wooteco.subway.AcceptanceTest;
+import wooteco.subway.acceptance.AcceptanceTest;
 
 import static org.assertj.core.api.Assertions.assertThat;
