+++ src/test/java/wooteco/subway/domain/station/StationRepositoryTest.java
@@ -4,10 +4,12 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.data.jdbc.DataJdbcTest;
 import org.springframework.data.relational.core.conversion.DbActionExecutionException;
+import org.springframework.test.context.jdbc.Sql;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 @DataJdbcTest
+@Sql("/truncate.sql")
 public class StationRepositoryTest {
     @Autowired
     private StationRepository stationRepository;
